divide go type declarations across separate flles
